refactor(database): number from small to integer

+++ database/migrations/2020_03_01_000110_create_box_volumes_table.php
@@ -19,7 +19,7 @@ return new class extends Migration {
         Schema::create('box_volumes', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('box_id');
-            $table->unsignedSmallInteger('number');
+            $table->unsignedInteger('number');
             $table->timestamps();
